backend/requirements.txt
# Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
python-dotenv>=1.0.0

# LLM (DeepSeek via OpenAI-compatible SDK)
openai>=1.50.0

# Embeddings & vector store
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0

# Stock data
yfinance>=0.2.40

# Computation
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.5.0